python/requirements.txt
Faker==0.8.15
grpcio==1.12.0
grpcio-tools==1.12.0
protobuf==3.5.2.post1
python-dateutil==2.7.3
six==1.11.0
text-unidecode==1.2
